pyproject.toml
[tool.poetry]
name = "quizhero-api-lambda-get-quizzes"
version = "0.1.0"
description = ""
authors = ["MatheusBLopes <matheusblopes10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
